+++ js/generic_lib.js
@@ -118,3 +118,80 @@ function jsJOMlib__getParameterByName(url, name)
 function jsJOMlib__isNumber(n) {
   return !isNaN(parseFloat(n)) && isFinite(n);
 }
+
+
+
+function jsJOMlib__date_formatted(format, splitter, date)
+{
+    if ( date==undefined ) date = new Date();
+    if ( toString.call(date) != "[object Date]" ) return false;
+
+    var day   = PHPjs_str_pad( date.getDate(),     2, '0', 'STR_PAD_LEFT' );
+    var month = PHPjs_str_pad( (date.getMonth() + 1),    2, '0', 'STR_PAD_LEFT' );
+    var year  = PHPjs_str_pad( date.getFullYear(), 4, '0', 'STR_PAD_LEFT' );
+
+    var retval = "";
+
+    var split = format.split(splitter);
+
+    // first part
+    for ( var i = 0 ; i < split.length ; i++ )
+    {
+        if ( split[i]=='dd' )   retval += day;
+        else
+        if ( split[i]=='mm' )   retval += month;
+        else
+        if ( split[i]=='yyyy' ) retval += year;
+
+        if ( i != (split.length - 1) ) retval += splitter;
+    }
+
+    return retval;
+}
+
+
+function PHPjs_str_pad (input, pad_length, pad_string, pad_type) {
+  // http://kevin.vanzonneveld.net
+  // +   original by: Kevin van Zonneveld (http://kevin.vanzonneveld.net)
+  // + namespaced by: Michael White (http://getsprink.com)
+  // +      input by: Marco van Oort
+  // +   bugfixed by: Brett Zamir (http://brett-zamir.me)
+  // *     example 1: str_pad('Kevin van Zonneveld', 30, '-=', 'STR_PAD_LEFT');
+  // *     returns 1: '-=-=-=-=-=-Kevin van Zonneveld'
+  // *     example 2: str_pad('Kevin van Zonneveld', 30, '-', 'STR_PAD_BOTH');
+  // *     returns 2: '------Kevin van Zonneveld-----'
+  var half = '',
+    pad_to_go;
+
+  var str_pad_repeater = function (s, len) {
+    var collect = '',
+      i;
+
+    while (collect.length < len) {
+      collect += s;
+    }
+    collect = collect.substr(0, len);
+
+    return collect;
+  };
+
+  input += '';
+  pad_string = pad_string !== undefined ? pad_string : ' ';
+
+  if (pad_type != 'STR_PAD_LEFT' && pad_type != 'STR_PAD_RIGHT' && pad_type != 'STR_PAD_BOTH') {
+    pad_type = 'STR_PAD_RIGHT';
+  }
+  if ((pad_to_go = pad_length - input.length) > 0) {
+    if (pad_type == 'STR_PAD_LEFT') {
+      input = str_pad_repeater(pad_string, pad_to_go) + input;
+    } else if (pad_type == 'STR_PAD_RIGHT') {
+      input = input + str_pad_repeater(pad_string, pad_to_go);
+    } else if (pad_type == 'STR_PAD_BOTH') {
+      half = str_pad_repeater(pad_string, Math.ceil(pad_to_go / 2));
+      input = half + input + half;
+      input = input.substr(0, pad_length);
+    }
+  }
+
+  return input;
+}